restructure testing framework
- each testing/projects/{project} is dynamically imported
    - provides definitions for expected class / file hierarchies
    - hierarchies are deep copied, tests that reuse previously
      mutated these
    - docs updated to use automodule, except `cpp with spaces`
      (not possible to document in sphinx)
- ignore `from collections import MutableMapping` warnings
  (see also #119)
- skip mac builds on CI until #122 is solved
- install correct beautiful soup package in tox.ini
- fix url parsing in docs/testproject.py
