+++ .github/workflows/test_python.yaml
@@ -14,16 +14,20 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [macos-latest, ubuntu-latest, windows-latest]
+        # os: [macos-latest, ubuntu-latest, windows-latest]
+        # TODO: either obtain doxygen 1.8.20 on mac, or find out why everything
+        # breaks in doxygen 1.9.2.
+        os: [ubuntu-latest, windows-latest]
     steps:
       - uses: actions/checkout@v2
       ##########################################################################
       - name: Install Doxygen (macOS)
         if: contains(matrix.os, 'macos')
         run: |
-          brew tap-new $USER/local-doxygen
-          brew extract --version=1.8.20 doxygen $USER/local-doxygen
-          HOMEBREW_NO_AUTO_UPDATE=1 brew install -v doxygen@1.8.20
+          # brew tap-new $USER/local-doxygen
+          # brew extract --version=1.8.20 doxygen $USER/local-doxygen
+          # HOMEBREW_NO_AUTO_UPDATE=1 brew install -v doxygen@1.8.20
+          HOMEBREW_NO_AUTO_UPDATE=1 brew install doxygen
       - name: Install Doxygen (Ubuntu)
         if: contains(matrix.os, 'ubuntu')
         run: |
