+++ docs/testing/projects.rst
@@ -1,0 +1,49 @@
+Testing Projects Module
+========================================================================================
+
+.. automodule:: testing.projects
+   :members:
+
+`` testing.projects.c_maths`` Project
+----------------------------------------------------------------------------------------
+
+.. automodule:: testing.projects.c_maths
+   :members:
+
+`` testing.projects.cpp with spaces`` Project
+----------------------------------------------------------------------------------------
+
+.. module:: testing.projects.cpp_with_spaces
+
+This cannot be documented because it has spaces in the name and autodoc cannot
+complete its import.
+
+`` testing.projects.cpp_fortran_mixed`` Project
+----------------------------------------------------------------------------------------
+
+.. automodule:: testing.projects.cpp_fortran_mixed
+   :members:
+
+`` testing.projects.cpp_func_overloads`` Project
+----------------------------------------------------------------------------------------
+
+.. automodule:: testing.projects.cpp_func_overloads
+   :members:
+
+`` testing.projects.cpp_long_names`` Project
+----------------------------------------------------------------------------------------
+
+.. automodule:: testing.projects.cpp_long_names
+   :members:
+
+`` testing.projects.cpp_nesting`` Project
+----------------------------------------------------------------------------------------
+
+.. automodule:: testing.projects.cpp_nesting
+   :members:
+
+`` testing.projects.cpp_pimpl`` Project
+----------------------------------------------------------------------------------------
+
+.. automodule:: testing.projects.cpp_pimpl
+   :members:
