+++ setup.cfg
@@ -46,6 +46,8 @@ filterwarnings =
     #   RemovedInSphinx30Warning: function based directive support is now deprecated.
     #                             Use class based directive instead.
     ignore::PendingDeprecationWarning:sphinx.util.docutils
+    # Any import of collections triggers this.
+    ignore: Using or importing the ABCs from 'collections' instead of from 'collections.abc' is deprecated since Python 3.3,and in 3.9 it will stop working
 
 [coverage:run]
 data_file = .coverage
