+++ testing/projects/__init__.py
@@ -1,0 +1,3 @@
+"""
+The test projects used to evaluate exhale.
+"""
