+++ tox.ini
@@ -12,15 +12,15 @@ skip_install = true
 deps =
     sphinx{env:SPHINX_VERSION:}
     breathe{env:BREATHE_VERSION:}
-    bs4
+    beautifulsoup4
     lxml
     six
 commands =
     {envpython} -m pip install -q -r requirements-dev.txt
+    {envpython} -c 'import sphinx; print("\033[36;1mSphinx version: %s\033[0m" % sphinx.__version__)'
+    {envpython} -c 'import breathe; print("\033[36;1mBreathe version: %s\033[0m" % breathe.__version__)'
+    {envpython} -c 'import sys; print("\033[36;1mPython version: %d.%d.%d\033[0m" % sys.version_info[0:3])'
     pytest . {posargs}
-    {envpython} -c 'import sphinx; print("Tested against Sphinx version %s." % sphinx.__version__)'
-    {envpython} -c 'import breathe; print("Tested against Breathe version %s." % breathe.__version__)'
-    {envpython} -c 'import sys; print("Tested against Python version %d.%d.%d." % sys.version_info[0:3])'
 
 [testenv:flake8]
 skip_install = true
